fix(core): settle the turmoil-feature clippy debts in the shard handlers
The script/function handlers grew past clippy's argument bound when the
write-admission seam added its `admission` parameter; allow it the same
way their `run_script`/`worker` siblings already do. The wait queue's
`RegisterRefused` is deliberately large — the refusal hands the entry
back so the caller keeps the response channel (FM-BLOCKING-006) — so its
three constructors take the same `result_large_err` allowance used
elsewhere for by-design large errors. Drop a same-type `as u32` cast in
the min-replicas gate.

+++ frogdb-server/crates/core/src/write_seam.rs
@@ -391,7 +391,7 @@ impl ShardWriteSeam {
             .as_ref()
             .map(|t| t.count_good_replicas(admission.min_replicas_max_lag))
             .unwrap_or(0);
-        if (good as u32) < admission.min_replicas {
+        if good < admission.min_replicas {
             return Err(NO_REPLICAS_ERR.to_string());
         }
         Ok(())
